+++ docs/examples/banneroverlay/desktop.tsx
@@ -1,13 +1,5 @@
 import { useState } from 'react';
-import {
-  Avatar,
-  BannerOverlay,
-  Button,
-  DeviceTypeProvider,
-  FixedZIndex,
-  Link,
-  Text,
-} from 'gestalt';
+import { BannerOverlay, Button, DeviceTypeProvider, FixedZIndex, Icon, Link, Text } from 'gestalt';
 
 export default function Example() {
   const [showComponent, setShowComponent] = useState(true);
@@ -44,7 +36,7 @@ export default function Example() {
           accessibilityLabel: 'Get the app',
         }}
         thumbnail={{
-          avatar: <Avatar name="Keerthi" src="https://i.ibb.co/ZfCZrY8/keerthi.jpg" />,
+          icon: <Icon accessibilityLabel="" color="brandPrimary" icon="pinterest" />,
         }}
         title="More to Explore"
         zIndex={new FixedZIndex(100)}