+++ packages/gestalt/src/Dropdown/Context.ts
@@ -3,15 +3,19 @@ import { Context, createContext } from 'react';
 type DropdownContextType = {
   id: string;
   hoveredItemIndex: number | null | undefined;
-  setHoveredItemIndex: (n: number) => void;
+  setHoveredItemIndex: (n: number | null | undefined) => void;
   setOptionRef: (arg1?: HTMLElement | null | undefined) => void;
+  focusedItemIndex: number | null | undefined;
+  setFocusedItemIndex: (n: number | null | undefined) => void;
 };
 
 const initialContextState = {
   id: '',
   hoveredItemIndex: -1,
   setHoveredItemIndex: () => {},
+  setFocusedItemIndex: () => {},
   setOptionRef: () => {},
+  focusedItemIndex: -1,
 } as const;
 
 const context: Context<DropdownContextType> =